feat: add a description to the embedded poll result
Now that the MVP is done it's time to consolidate and refactor.
The bot is broken in many ways and will not survive message deletion
or such shenanigans as reacting while it is adding its own reactions.

It's brittle and easily abused, but it does the job if not mishandled.
It's time to add issues and sweep the broom.